refactor: move `assembleSelectionPredicate` to `parseSelectionPredicate`
(as it's happening during parse)
