+++ test/compile/selection/predicate.test.ts
@@ -1,7 +1,6 @@
 import {nonPosition} from '../../../src/compile/mark/mixins';
 import {expression} from '../../../src/compile/predicate';
-import {assembleSelectionPredicate as predicate} from '../../../src/compile/selection/assemble';
-import {parseUnitSelection} from '../../../src/compile/selection/parse';
+import {parseSelectionPredicate as predicate, parseUnitSelection} from '../../../src/compile/selection/parse';
 import {parseUnitModel} from '../../util';
 
 describe('Selection Predicate', () => {
